test: move fuse test ATTRIBUTES to conftest.py as fixture
